+++ packages/storybook-vue/.storybook/usage-addon/usage.js
@@ -15,10 +15,10 @@ import { Description } from '@storybook/components';
 import { useStorybookState, useGlobals } from '@storybook/api';
 import accordion_en from 'raw-loader!../../stories/components/accordion/accordion.md';
 import accordion_de from 'raw-loader!../../stories/components/accordion/accordion_de.md';
-import brandHeader_en from 'raw-loader!../../stories/components/brand-header-navigation/brand-header.md';
-import brandHeader_de from 'raw-loader!../../stories/components/brand-header-navigation/brand-header_de.md';
-import betaBrandHeader_en from 'raw-loader!../../stories/components/beta-brand-header-navigation/beta-brand-header.md';
-import betaBrandHeader_de from 'raw-loader!../../stories/components/beta-brand-header-navigation/beta-brand-header_de.md';
+import brandHeader_en from 'raw-loader!../../stories/deprecated/brand-header-navigation/brand-header.md';
+import brandHeader_de from 'raw-loader!../../stories/deprecated/brand-header-navigation/brand-header_de.md';
+import telekomBrandHeader_en from 'raw-loader!../../stories/components/telekom-brand-header-navigation/telekom-brand-header.md';
+import telekomBrandHeader_de from 'raw-loader!../../stories/components/telekom-brand-header-navigation/telekom-brand-header_de.md';
 import breadcrumb_en from 'raw-loader!../../stories/components/breadcrumb/breadcrumb.md';
 import breadcrumb_de from 'raw-loader!../../stories/components/breadcrumb/breadcrumb_de.md';
 import button_en from 'raw-loader!../../stories/components/button/button.md';
@@ -91,14 +91,14 @@ import textArea_en from 'raw-loader!../../stories/components/text-area/text-area
 import textArea_de from 'raw-loader!../../stories/components/text-area/text-area_de.md';
 import textField_en from 'raw-loader!../../stories/components/text-field/text-field.md';
 import textField_de from 'raw-loader!../../stories/components/text-field/text-field_de.md';
-import footer_en from 'raw-loader!../../stories/components/footer/footer.md';
-import footer_de from 'raw-loader!../../stories/components/footer/footer_de.md';
-import betaFooter_en from 'raw-loader!../../stories/components/beta-footer/beta-footer.md';
-import betaFooter_de from 'raw-loader!../../stories/components/beta-footer/beta-footer_de.md';
+import footer_en from 'raw-loader!../../stories/deprecated/footer/footer.md';
+import footer_de from 'raw-loader!../../stories/deprecated/footer/footer_de.md';
+import telekomFooter_en from 'raw-loader!../../stories/components/telekom-footer/telekom-footer.md';
+import telekomFooter_de from 'raw-loader!../../stories/components/telekom-footer/telekom-footer_de.md';
 import callout_en from 'raw-loader!../../stories/components/callout/callout.md';
 import callout_de from 'raw-loader!../../stories/components/callout/callout_de.md';
-import toggleGroup_de from 'raw-loader!../../stories/components/toggle-group/toggle-group_de.md';
-import toggleGroup_en from 'raw-loader!../../stories/components/toggle-group/toggle-group.md';
+import toggleGroup_de from 'raw-loader!../../stories/deprecated/toggle-group/toggle-group_de.md';
+import toggleGroup_en from 'raw-loader!../../stories/deprecated/toggle-group/toggle-group.md';
 import tooltip_en from 'raw-loader!../../stories/components/tooltip/tooltip.md';
 import tooltip_de from 'raw-loader!../../stories/components/tooltip/tooltip_de.md';
 
@@ -127,8 +127,8 @@ const Usage = (props) => {
     accordion_de,
     'brand-header-navigation_en': brandHeader_en,
     'brand-header-navigation_de': brandHeader_de,
-    'telekom-brand-header-navigation_en': betaBrandHeader_en,
-    'telekom-brand-header-navigation_de': betaBrandHeader_de,
+    'telekom-brand-header-navigation_en': telekomBrandHeader_en,
+    'telekom-brand-header-navigation_de': telekomBrandHeader_de,
     breadcrumb_en,
     breadcrumb_de,
     button_en,
@@ -207,8 +207,8 @@ const Usage = (props) => {
     'text-field_de': textField_de,
     footer_en,
     footer_de,
-    'telekom-footer_en': betaFooter_en,
-    'telekom-footer_de': betaFooter_de,    
+    'telekom-footer_en': telekomFooter_en,
+    'telekom-footer_de': telekomFooter_de,
     'toggle-group_en': toggleGroup_en,
     'toggle-group_de': toggleGroup_de,
     tooltip_en,