+++ packages/storybook-vue/stories/components/telekom-brand-header-navigation/SlotMainNav.vue
@@ -1,0 +1,19 @@
+<template>
+  <scale-telekom-nav-list variant="main-nav" slot="main-nav" aria-label="Main navigation links">
+    <scale-telekom-nav-item active aria-label="Topic One">
+      <a href="#">Topic One</a>
+    </scale-telekom-nav-item>
+    <scale-telekom-nav-item aria-label="Topic Two">
+      <a href="#">Topic Two</a>
+    </scale-telekom-nav-item>
+    <scale-telekom-nav-item aria-label="Topic Three">
+      <a href="#">Topic Three</a>
+    </scale-telekom-nav-item>
+    <scale-telekom-nav-item aria-label="Topic Four">
+      <a href="#">Topic Four</a>
+    </scale-telekom-nav-item>
+    <scale-telekom-nav-item aria-label="Topic Five">
+      <a href="#">Topic Five</a>
+    </scale-telekom-nav-item>
+  </scale-telekom-nav-list>
+</template>